+++ main.py
@@ -2,10 +2,10 @@ from fastapi import FastAPI, routing
 from src.db.database_connection import engine, SessionLocal
 from src.db import models, crud
 from src.routers import (
-    language_processing,
-    text_to_voice,
-    voice_to_text,
-    image_generation,
+    language_processing_routes,
+    text_to_voice_routes,
+    voice_to_text_routes,
+    image_generation_routes,
 )
 from src.security import security_routes, handler, security_schemas
 
@@ -47,7 +47,7 @@ async def startup_event():
 
 
 app.include_router(security_routes.router)
-app.include_router(language_processing.router)
-app.include_router(text_to_voice.router)
-app.include_router(voice_to_text.router)
-app.include_router(image_generation.router)
+app.include_router(language_processing_routes.router)
+app.include_router(text_to_voice_routes.router)
+app.include_router(voice_to_text_routes.router)
+app.include_router(image_generation_routes.router)
