+++ src/ai_services/image_generation_services.py
