+++ src/routers/image_generation_routes.py
@@ -2,8 +2,8 @@ from fastapi import APIRouter, Security
 from typing import Union, Annotated
 
 import os
-from ..ai_services.image_generation import Dalle
-from ..schemas.image_generation import ImageData
+from ..ai_services.image_generation_services import Dalle
+from ..schemas.image_generation_schemas import ImageData
 
 from ..security.handler import get_current_active_user
