+++ src/schemas/image_generation_schemas.py
